# Use abstract command to access registers if possible