This is the only way that the spec requires that GPRs are accessed. This gets us a lot closer to debugging cores which don't have a program buffer.

As a side benefit, it's a bit faster too (WriteGprs improved from 29s to 19s on spike, and from 17s to 15s on an FPGA target).

Some smaller changes:
* When an abstract command is executed and fails, immediately clear cmderr while returning the error.
* Also change debug code that decodes DMI values so it's easier to maintain.